ModelTraining/requirements.txt
# データ処理
pandas>=1.5.0
numpy>=1.21.0
scikit-learn>=1.0.0

# 深層学習
tensorflow>=2.10.0
keras>=2.10.0

# データ取得
yfinance>=0.2.0
websocket-client>=1.3.0
requests>=2.28.0

# データ可視化
matplotlib>=3.5.0
seaborn>=0.12.0
plotly>=5.10.0

# テクニカル指標
ta-lib>=0.4.24

# 統計・最適化
scipy>=1.7.0
statsmodels>=0.13.0

# XGBoost
xgboost>=1.5.0

# ユーティリティ
python-dotenv>=0.20.0
joblib>=1.1.0

# 型ヒント
typing-extensions>=4.0.0

# ログ
logging>=0.5.0

# JSON処理
jsonschema>=4.0.0